Share the metadata_controller test fixture via a package conftest
